fix(S-03): extract NavigationMapper to separate mapping from navigation (SRP)
NavigationService had three reasons to change: navigation tree structure,
domain-to-DTO mapping, and sort ordering. At 299 lines with 15 private
methods, any change to DTO shapes, tree rendering, sort rules, or
credit/transaction mapping all modified the same class.

NavigationMapper (internal static) now owns:
- BuildBrokerTreeNode / BuildPortfolioTreeNode / BuildAssetTreeNode
- MapBroker / MapPortfolio / MapAsset / MapTransaction / MapCredit
- CalculateTotals
- OrderByNameWithEncerradasLast / IsEncerradas

NavigationService retains only the repository interaction and delegates all
mapping and tree-building to NavigationMapper. The class drops from 299 to
115 lines, each with a single reason to change.
