+++ Financial.Application/Services/NavigationService.cs
@@ -9,7 +9,6 @@ namespace Financial.Application.Services;
 
 public sealed class NavigationService : INavigationService
 {
-    private const string EncerradasName = "Encerradas";
     private readonly IRepository _repository;
 
     public NavigationService(IRepository repository)
@@ -29,7 +28,7 @@ public sealed class NavigationService : INavigationService
 
         foreach (var broker in brokers)
         {
-            rootNode.Children.Add(BuildBrokerTreeNode(broker));
+            rootNode.Children.Add(NavigationMapper.BuildBrokerTreeNode(broker));
         }
 
         return rootNode;
@@ -44,8 +43,8 @@ public sealed class NavigationService : INavigationService
             return null;
         }
 
-        var assets = _repository.GetAssetsByBrokerPortfolio(brokerName, portfolioName);
-        var asset = assets.FirstOrDefault(a => a.Name == assetName);
+        var asset = _repository.GetAssetsByBrokerPortfolio(brokerName, portfolioName)
+            .FirstOrDefault(a => a.Name == assetName);
 
         if (asset == null)
         {
@@ -53,16 +52,16 @@ public sealed class NavigationService : INavigationService
         }
 
         var transactions = asset.Transactions
-            .Select(MapTransaction)
+            .Select(NavigationMapper.MapTransaction)
             .OrderByDescending(t => t.Date)
             .ToList();
 
         var credits = asset.Credits
-            .Select(MapCredit)
+            .Select(NavigationMapper.MapCredit)
             .OrderByDescending(c => c.Date)
             .ToList();
 
-        var (totalBought, totalSold, totalCredits) = CalculateTotals(asset);
+        var (totalBought, totalSold, totalCredits) = NavigationMapper.CalculateTotals(asset);
 
         return new AssetDetailsDTO
         {
@@ -88,8 +87,8 @@ public sealed class NavigationService : INavigationService
 
     public IEnumerable<BrokerNodeDTO> GetBrokers()
     {
-        var brokers = OrderByNameWithEncerradasLast(_repository.GetBrokerList(), broker => broker.Name);
-        return brokers.Select(MapBroker).ToList();
+        var brokers = NavigationMapper.OrderByNameWithEncerradasLast(_repository.GetBrokerList(), b => b.Name);
+        return brokers.Select(NavigationMapper.MapBroker).ToList();
     }
 
     public IReadOnlyList<CreditDTO> GetCreditsByBroker(string brokerName)
@@ -102,7 +101,7 @@ public sealed class NavigationService : INavigationService
         return _repository.GetAssetsByBroker(brokerName)
             .Where(asset => asset.Active)
             .SelectMany(asset => asset.Credits)
-            .Select(MapCredit)
+            .Select(NavigationMapper.MapCredit)
             .OrderByDescending(credit => credit.Date)
             .ToList();
     }
@@ -117,183 +116,8 @@ public sealed class NavigationService : INavigationService
         return _repository.GetAssetsByBrokerPortfolio(brokerName, portfolioName)
             .Where(asset => asset.Active)
             .SelectMany(asset => asset.Credits)
-            .Select(MapCredit)
+            .Select(NavigationMapper.MapCredit)
             .OrderByDescending(credit => credit.Date)
             .ToList();
-    }
-
-    private static TreeNodeDTO BuildBrokerTreeNode(BrokerNodeDTO broker)
-    {
-        var brokerNode = new TreeNodeDTO
-        {
-            NodeType = TreeNodeTypes.Broker,
-            DisplayName = $"{broker.Name} ({broker.Currency})",
-            Metadata = new Dictionary<string, object>
-            {
-                ["BrokerName"] = broker.Name,
-                ["Currency"] = broker.Currency,
-                ["PortfolioCount"] = broker.PortfolioCount,
-                ["TotalAssets"] = broker.TotalAssets
-            }
-        };
-
-        foreach (var portfolio in broker.Portfolios)
-        {
-            brokerNode.Children.Add(BuildPortfolioTreeNode(portfolio));
-        }
-
-        return brokerNode;
-    }
-
-    private static TreeNodeDTO BuildPortfolioTreeNode(PortfolioNodeDTO portfolio)
-    {
-        var portfolioNode = new TreeNodeDTO
-        {
-            NodeType = TreeNodeTypes.Portfolio,
-            DisplayName = $"{portfolio.Name} ({portfolio.AssetCount} assets)",
-            Metadata = new Dictionary<string, object>
-            {
-                ["PortfolioName"] = portfolio.Name,
-                ["AssetCount"] = portfolio.AssetCount,
-                ["ActiveAssetCount"] = portfolio.ActiveAssetCount
-            }
-        };
-
-        foreach (var asset in portfolio.Assets)
-        {
-            portfolioNode.Children.Add(BuildAssetTreeNode(asset));
-        }
-
-        return portfolioNode;
-    }
-
-    private static TreeNodeDTO BuildAssetTreeNode(AssetNodeDTO asset)
-    {
-        return new TreeNodeDTO
-        {
-            NodeType = TreeNodeTypes.Asset,
-            DisplayName = asset.Name,
-            Metadata = new Dictionary<string, object>
-            {
-                ["AssetName"] = asset.Name,
-                ["Ticker"] = asset.Ticker,
-                ["Exchange"] = asset.Exchange,
-                ["ISIN"] = asset.ISIN,
-                ["Country"] = asset.Country,
-                ["LocalTypeCode"] = asset.LocalTypeCode,
-                ["GlobalAssetClass"] = asset.Class,
-                ["Quantity"] = asset.Quantity,
-                ["AveragePrice"] = asset.AveragePrice,
-                ["IsActive"] = asset.IsActive,
-                ["TransactionCount"] = asset.TransactionCount,
-                ["CreditCount"] = asset.CreditCount
-            }
-        };
-    }
-
-    private static BrokerNodeDTO MapBroker(Broker broker)
-    {
-        return new BrokerNodeDTO
-        {
-            Name = broker.Name,
-            Currency = broker.Currency,
-            PortfolioCount = broker.Portfolios.Count,
-            TotalAssets = broker.Portfolios.Sum(p => p.Assets.Count),
-            Portfolios = MapPortfolios(broker.Portfolios).ToList()
-        };
-    }
-
-    private static IEnumerable<PortfolioNodeDTO> MapPortfolios(IEnumerable<Portfolio> portfolios)
-    {
-        return OrderByNameWithEncerradasLast(portfolios, portfolio => portfolio.Name)
-            .Select(MapPortfolio);
-    }
-
-    private static PortfolioNodeDTO MapPortfolio(Portfolio portfolio)
-    {
-        return new PortfolioNodeDTO
-        {
-            Name = portfolio.Name,
-            AssetCount = portfolio.Assets.Count,
-            ActiveAssetCount = portfolio.Assets.Count(a => a.Active),
-            Assets = MapAssets(portfolio.Assets).ToList()
-        };
-    }
-
-    private static IEnumerable<AssetNodeDTO> MapAssets(IEnumerable<Asset> assets)
-    {
-        return OrderByNameWithEncerradasLast(assets, asset => asset.Name)
-            .Select(MapAsset);
-    }
-
-    private static AssetNodeDTO MapAsset(Asset asset)
-    {
-        return new AssetNodeDTO
-        {
-            Name = asset.Name,
-            Ticker = asset.Ticker,
-            Exchange = asset.Exchange,
-            ISIN = asset.ISIN,
-            Country = asset.Country,
-            LocalTypeCode = asset.LocalTypeCode,
-            Class = asset.Class,
-            Quantity = asset.Quantity,
-            AveragePrice = asset.AveragePrice,
-            IsActive = asset.Active,
-            TransactionCount = asset.Transactions.Count,
-            CreditCount = asset.Credits.Count
-        };
-    }
-
-    private static TransactionDTO MapTransaction(Transaction transaction)
-    {
-        return new TransactionDTO
-        {
-            Id = transaction.Id,
-            Date = transaction.Date,
-            Type = transaction.Type.ToString(),
-            Quantity = transaction.Quantity,
-            UnitPrice = transaction.UnitPrice,
-            Fees = transaction.Fees,
-            TotalPrice = transaction.TotalPrice
-        };
-    }
-
-    private static CreditDTO MapCredit(Credit credit)
-    {
-        return new CreditDTO
-        {
-            Id = credit.Id,
-            Date = credit.Date,
-            Type = credit.Type.ToString(),
-            Value = credit.Value
-        };
-    }
-
-    private static (decimal TotalBought, decimal TotalSold, decimal TotalCredits) CalculateTotals(Asset asset)
-    {
-        var totalBought = asset.Transactions
-            .Where(t => t.Type == Transaction.TransactionType.Buy)
-            .Sum(t => t.TotalPrice);
-
-        var totalSold = asset.Transactions
-            .Where(t => t.Type == Transaction.TransactionType.Sell)
-            .Sum(t => t.TotalPrice);
-
-        var totalCredits = asset.Credits.Sum(c => c.Value);
-
-        return (totalBought, totalSold, totalCredits);
-    }
-
-    private static IEnumerable<T> OrderByNameWithEncerradasLast<T>(IEnumerable<T> source, Func<T, string> nameSelector)
-    {
-        return source
-            .OrderBy(item => IsEncerradas(nameSelector(item)) ? 1 : 0)
-            .ThenBy(item => nameSelector(item) ?? string.Empty, StringComparer.CurrentCultureIgnoreCase);
-    }
-
-    private static bool IsEncerradas(string? name)
-    {
-        return string.Equals(name?.Trim(), EncerradasName, StringComparison.CurrentCultureIgnoreCase);
     }
 }
